pyproject.toml
[tool.poetry]
name = "schnephro"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.2.0"
mysqlclient = "^2.2.0"
django-environ = "^0.11.2"
dj-rest-auth = "^5.0.1"
djangorestframework-simplejwt = "^5.3.0"
django-allauth = "^0.57.0"
xhtml2pdf = "^0.2.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
